Add comprehensive documentation for Surge language specification and pragma directives
- Moved few documents: LANGUAGE.md, PARALLEL_COMPILATION.md, PRAGMA.md, SEMANTICS_PLAN.md, and TRACING.md, detailing the Surge language features, parallel compilation capabilities, pragma directives, semantic analysis plans, and tracing system.
- The LANGUAGE.md document outlines the language design goals, implementation snapshots, and lexical structure, providing a clear specification for developers.
- The PRAGMA.md document explains the pragma mechanism for declaring module properties and compilation modes, enhancing clarity on its usage.
- The SEMANTICS_PLAN.md document details the planned phases for semantic analysis, including symbol resolution and type checking, ensuring a structured approach to implementation.
- The TRACING.md document describes the tracing system for debugging and performance profiling, including usage examples and trace levels, improving developer experience and diagnostics.
- Enhanced overall documentation to support better understanding and usage of the Surge compiler and its features.
