+++ internal/sema/contract_match.go
@@ -190,173 +190,6 @@ func (tc *typeChecker) checkContractSatisfaction(target types.TypeID, bound symb
 	return ok
 }
 
-func (tc *typeChecker) validateFunctionCall(sym *symbols.Symbol, call *ast.ExprCallData, argTypes []types.TypeID) {
-	if sym == nil || call == nil || tc.builder == nil {
-		return
-	}
-	fnItem, ok := tc.builder.Items.Fn(sym.Decl.Item)
-	if !ok || fnItem == nil {
-		return
-	}
-	bindings := tc.inferTypeParamBindings(sym, fnItem, argTypes, call)
-	if len(sym.TypeParamSymbols) > 0 {
-		tc.enforceContractBounds(sym.TypeParamSymbols, bindings, tc.exprSpan(call.Target))
-	}
-}
-
-func (tc *typeChecker) inferTypeParamBindings(sym *symbols.Symbol, fn *ast.FnItem, argTypes []types.TypeID, call *ast.ExprCallData) map[source.StringID]bindingInfo {
-	if sym == nil || fn == nil || len(sym.TypeParams) == 0 || tc.builder == nil || call == nil {
-		return nil
-	}
-	result := make(map[source.StringID]bindingInfo, len(sym.TypeParams))
-	indexByName := make(map[source.StringID]struct{}, len(sym.TypeParams))
-	for _, name := range sym.TypeParams {
-		indexByName[name] = struct{}{}
-	}
-	paramIDs := tc.builder.Items.GetFnParamIDs(fn)
-	for i, pid := range paramIDs {
-		if i >= len(argTypes) {
-			break
-		}
-		argType := argTypes[i]
-		if argType == types.NoTypeID {
-			continue
-		}
-		argSpan := tc.exprSpan(call.Args[i])
-		argSym := tc.symbolForExpr(call.Args[i])
-		argValType := tc.valueType(argType)
-		if argSym.IsValid() {
-			if boundType := tc.bindingType(argSym); boundType != types.NoTypeID {
-				argValType = boundType
-			}
-		}
-		param := tc.builder.Items.FnParam(pid)
-		if param == nil {
-			continue
-		}
-		if name := tc.paramTypeParamName(param.Type, indexByName); name != source.NoStringID {
-			result[name] = bindingInfo{typ: argValType, span: argSpan, sym: argSym}
-		}
-	}
-	return result
-}
-
-func (tc *typeChecker) paramTypeParamName(typeID ast.TypeID, allowed map[source.StringID]struct{}) source.StringID {
-	if typeID == ast.NoTypeID || tc.builder == nil {
-		return source.NoStringID
-	}
-	expr := tc.builder.Types.Get(typeID)
-	if expr == nil || expr.Kind != ast.TypeExprPath {
-		return source.NoStringID
-	}
-	path, ok := tc.builder.Types.Path(typeID)
-	if !ok || path == nil || len(path.Segments) != 1 {
-		return source.NoStringID
-	}
-	seg := path.Segments[0]
-	if len(seg.Generics) > 0 {
-		return source.NoStringID
-	}
-	if _, ok := allowed[seg.Name]; ok {
-		return seg.Name
-	}
-	return source.NoStringID
-}
-
-func (tc *typeChecker) enforceContractBounds(params []symbols.TypeParamSymbol, bindings map[source.StringID]bindingInfo, span source.Span) {
-	if len(params) == 0 || tc.reporter == nil {
-		return
-	}
-	for _, param := range params {
-		binding := bindings[param.Name]
-		concrete := binding.typ
-		if concrete == types.NoTypeID {
-			continue
-		}
-		reportSpan := binding.span
-		if reportSpan == (source.Span{}) {
-			reportSpan = span
-		}
-		typeLabel := tc.bindingTypeLabel(binding)
-		for _, bound := range param.Bounds {
-			inst := bound
-			inst.GenericArgs = tc.substituteBoundArgs(bound.GenericArgs, bindings)
-			if tc.typeParamSatisfiesBound(concrete, inst, bindings) {
-				continue
-			}
-			tc.checkContractSatisfaction(concrete, inst, reportSpan, typeLabel)
-		}
-	}
-}
-
-func (tc *typeChecker) substituteBoundArgs(args []types.TypeID, bindings map[source.StringID]bindingInfo) []types.TypeID {
-	if len(args) == 0 {
-		return nil
-	}
-	out := make([]types.TypeID, len(args))
-	for i, arg := range args {
-		out[i] = tc.substituteTypeParamByName(arg, bindings)
-	}
-	return out
-}
-
-func (tc *typeChecker) substituteTypeParamByName(id types.TypeID, bindings map[source.StringID]bindingInfo) types.TypeID {
-	if id == types.NoTypeID || tc.types == nil {
-		return id
-	}
-	resolved := tc.resolveAlias(id)
-	tt, ok := tc.types.Lookup(resolved)
-	if !ok {
-		return resolved
-	}
-	if tt.Kind == types.KindGenericParam {
-		name := tc.typeParamNames[resolved]
-		if name == source.NoStringID {
-			if info, okInfo := tc.types.TypeParamInfo(resolved); okInfo && info != nil {
-				name = info.Name
-				if name != source.NoStringID {
-					tc.typeParamNames[resolved] = name
-				}
-			}
-		}
-		if name != source.NoStringID {
-			if concrete := bindings[name].typ; concrete != types.NoTypeID {
-				return concrete
-			}
-		}
-		return resolved
-	}
-	if tt.Kind == types.KindStruct {
-		if elem, ok := tc.arrayElemType(resolved); ok {
-			inner := tc.substituteTypeParamByName(elem, bindings)
-			if inner == elem {
-				return resolved
-			}
-			return tc.instantiateArrayType(inner)
-		}
-	}
-	switch tt.Kind {
-	case types.KindPointer, types.KindReference, types.KindOwn:
-		elem := tc.substituteTypeParamByName(tt.Elem, bindings)
-		if elem == tt.Elem {
-			return resolved
-		}
-		clone := tt
-		clone.Elem = elem
-		return tc.types.Intern(clone)
-	case types.KindArray:
-		elem := tc.substituteTypeParamByName(tt.Elem, bindings)
-		if elem == tt.Elem {
-			return resolved
-		}
-		clone := tt
-		clone.Elem = elem
-		return tc.types.Intern(clone)
-	default:
-		return resolved
-	}
-}
-
 func (tc *typeChecker) contractRequirementSet(contractDecl *ast.ContractDecl, scope symbols.ScopeID) (contractRequirements, bool) {
 	reqs := contractRequirements{
 		fields:     make(map[source.StringID]types.TypeID),
@@ -503,215 +336,4 @@ func (tc *typeChecker) instantiateContractRequirements(sym *symbols.Symbol, spec
 		reqs.methods[mname] = methods
 	}
 	return reqs
-}
-
-func (tc *typeChecker) collectTypeFields(target types.TypeID) map[source.StringID]types.TypeID {
-	fields := make(map[source.StringID]types.TypeID)
-	target = tc.valueType(target)
-	if target == types.NoTypeID {
-		return fields
-	}
-	if info, _ := tc.structInfoForType(target); info != nil {
-		for _, field := range info.Fields {
-			fields[field.Name] = field.Type
-		}
-	}
-	for _, field := range tc.externFieldsForType(target) {
-		if _, exists := fields[field.Name]; !exists {
-			fields[field.Name] = field.Type
-		}
-	}
-	return fields
-}
-
-func (tc *typeChecker) collectFieldAttrs(target types.TypeID) map[source.StringID][]source.StringID {
-	attrMap := make(map[source.StringID][]source.StringID)
-	target = tc.valueType(target)
-	if target == types.NoTypeID {
-		return attrMap
-	}
-	if info, _ := tc.structInfoForType(target); info != nil {
-		for _, field := range info.Fields {
-			attrMap[field.Name] = field.Attrs
-		}
-	}
-	for _, field := range tc.externFieldsForType(target) {
-		if _, exists := attrMap[field.Name]; !exists {
-			attrMap[field.Name] = field.Attrs
-		}
-	}
-	return attrMap
-}
-
-// returns 1 if satisfied, 0 if signature mismatch, -1 if missing entirely, -2 if only attrs/modifiers mismatch
-func (tc *typeChecker) ensureMethodSatisfies(target types.TypeID, name source.StringID, req *methodRequirement, reportSpan source.Span, contractName string) int {
-	if req == nil {
-		return 0
-	}
-	if len(req.params) > 0 && !tc.contractTypesEqual(req.params[0], target) {
-		tc.report(diag.SemaContractSelfType, reportSpan, "type %s method '%s' must have self %s per contract %s, got %s", tc.contractTypeLabel(target), tc.lookupName(name), tc.typeLabel(target), contractName, tc.typeLabel(req.params[0]))
-		return 0
-	}
-
-	actual := tc.methodsForType(target, name)
-	if len(actual) == 0 {
-		return -1
-	}
-
-	attrMismatch := false
-	for _, cand := range actual {
-		var aligned methodRequirement
-		switch {
-		case len(cand.params) == len(req.params):
-			aligned = *req
-		case len(cand.params) == len(req.params)+1:
-			aligned = *req
-			aligned.params = append([]types.TypeID{target}, req.params...)
-		default:
-			continue
-		}
-		if len(aligned.params) > 0 && !tc.contractTypesEqual(aligned.params[0], target) {
-			continue
-		}
-		if match, attrBad := tc.contractSignatureMatches(&aligned, cand); match {
-			return 1
-		} else if attrBad {
-			attrMismatch = true
-		}
-	}
-	if attrMismatch {
-		return -2
-	}
-	return 0
-}
-
-func (tc *typeChecker) methodsForType(target types.TypeID, name source.StringID) []methodSignature {
-	// Трассировка поиска методов для типа
-	var span *trace.Span
-	if tc.tracer != nil && tc.tracer.Level() >= trace.LevelDebug {
-		span = trace.Begin(tc.tracer, trace.ScopeNode, "methods_for_type", 0)
-		span.WithExtra("type", tc.typeLabel(target))
-	}
-	defer func() {
-		if span != nil {
-			span.End("")
-		}
-	}()
-
-	if target == types.NoTypeID || name == source.NoStringID {
-		return nil
-	}
-	nameLiteral := tc.lookupName(name)
-	if nameLiteral == "" {
-		return nil
-	}
-
-	candidates := tc.typeKeyCandidates(target)
-	methods := make([]methodSignature, 0, 4)
-
-	for _, cand := range candidates {
-		if cand.key == "" {
-			continue
-		}
-		for _, sig := range tc.lookupMagicMethods(cand.key, nameLiteral) {
-			if sig == nil || len(sig.Params) == 0 || !typeKeyEqual(sig.Params[0], cand.key) {
-				continue
-			}
-			if ms, ok := tc.signatureToTypes(sig); ok {
-				ms.pub = false
-				ms.async = false
-				methods = append(methods, ms)
-			}
-		}
-	}
-
-	if tc.symbols != nil && tc.symbols.Table != nil && tc.symbols.Table.Symbols != nil {
-		if data := tc.symbols.Table.Symbols.Data(); data != nil {
-			for i := range data {
-				sym := &data[i]
-				if sym.Kind != symbols.SymbolFunction || sym.Signature == nil {
-					continue
-				}
-				if tc.symbolName(sym.Name) != nameLiteral {
-					continue
-				}
-				if len(sym.Signature.Params) == 0 {
-					continue
-				}
-				for _, cand := range candidates {
-					if cand.key == "" || !typeKeyEqual(sym.Signature.Params[0], cand.key) {
-						continue
-					}
-					if ms, ok := tc.signatureToTypes(sym.Signature); ok {
-						ms.pub = sym.Flags&symbols.SymbolFlagPublic != 0
-						if fn, okFn := tc.builder.Items.Fn(sym.Decl.Item); okFn && fn != nil {
-							ms.attrs = tc.attrNames(fn.AttrStart, fn.AttrCount)
-							ms.async = fn.Flags&ast.FnModifierAsync != 0
-						}
-						methods = append(methods, ms)
-					}
-					break
-				}
-			}
-		}
-	}
-
-	if span != nil {
-		span.WithExtra("found", fmt.Sprintf("%d", len(methods)))
-	}
-
-	return methods
-}
-
-func (tc *typeChecker) signatureToTypes(sig *symbols.FunctionSignature) (methodSignature, bool) {
-	ms := methodSignature{}
-	if sig == nil {
-		return ms, false
-	}
-	params := make([]types.TypeID, 0, len(sig.Params))
-	ok := true
-	for _, p := range sig.Params {
-		paramType := tc.typeFromKey(p)
-		params = append(params, paramType)
-		if paramType == types.NoTypeID {
-			ok = false
-		}
-	}
-	ms.params = params
-	ms.result = tc.typeFromKey(sig.Result)
-	if ms.result == types.NoTypeID {
-		ok = false
-	}
-	return ms, ok
-}
-
-func (tc *typeChecker) contractSignatureMatches(expected *methodRequirement, actual methodSignature) (match, attrMismatch bool) {
-	if expected == nil {
-		return false, false
-	}
-	if len(expected.params) != len(actual.params) {
-		return false, false
-	}
-	for i := range expected.params {
-		if !tc.contractTypesEqual(expected.params[i], actual.params[i]) {
-			return false, false
-		}
-	}
-	if !tc.contractTypesEqual(expected.result, actual.result) {
-		return false, false
-	}
-	if expected.pub != actual.pub || expected.async != actual.async {
-		return false, true
-	}
-	if !tc.attrSetsEqual(expected.attrs, actual.attrs) {
-		return false, true
-	}
-	return true, false
-}
-
-func (tc *typeChecker) contractTypesEqual(expected, actual types.TypeID) bool {
-	if expected == types.NoTypeID || actual == types.NoTypeID {
-		return false
-	}
-	return tc.resolveAlias(expected) == tc.resolveAlias(actual)
 }
